+++ src/Savvyio.Core/SavvyioOptionsExtensions.cs
@@ -2,6 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using System.Runtime.CompilerServices;
+using Cuemon;
+using Cuemon.Collections.Generic;
 using Cuemon.Extensions;
 using Cuemon.Extensions.Collections.Generic;
 using Savvyio.Dispatchers;
@@ -19,6 +22,9 @@ namespace Savvyio
         /// <param name="options">The <see cref="SavvyioOptions"/> to extend.</param>
         /// <param name="assemblies">The assemblies to scan for <see cref="IDispatcher"/> implementations.</param>
         /// <returns>A reference to <paramref name="options"/> so that additional configuration calls can be chained.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="assemblies"/> cannot be null.
+        /// </exception>
         public static SavvyioOptions AddDispatchers(this SavvyioOptions options, params Assembly[] assemblies)
         {
             return AddDependenciesCore<IDispatcher>(options, (service, implementation) => options.AddDispatcher(service, implementation), assemblies);
@@ -30,15 +36,20 @@ namespace Savvyio
         /// <param name="options">The <see cref="SavvyioOptions"/> to extend.</param>
         /// <param name="assemblies">The assemblies to scan for <see cref="IHandler"/> implementations.</param>
         /// <returns>A reference to <paramref name="options"/> so that additional configuration calls can be chained.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="assemblies"/> cannot be null.
+        /// </exception>
         public static SavvyioOptions AddHandlers(this SavvyioOptions options, params Assembly[] assemblies)
         {
             return AddDependenciesCore<IHandler>(options, (service, implementation) => options.AddHandler(service, implementation), assemblies);
         }
 
-        private static SavvyioOptions AddDependenciesCore<T>(SavvyioOptions options, Action<Type, Type> servicesManager, IEnumerable<Assembly> assembliesToScan)
+        private static SavvyioOptions AddDependenciesCore<T>(SavvyioOptions options, Action<Type, Type> servicesManager, IEnumerable<Assembly> assemblies)
         {
+            Validator.ThrowIfNull(assemblies, nameof(assemblies));
+
             var definedTypes = new List<TypeInfo>();
-            var assemblies = (assembliesToScan ?? AppDomain.CurrentDomain.GetAssemblies().Except(typeof(SavvyioOptions).Assembly.Yield())).ToList();
+
             foreach (var assembly in assemblies)
             {
                 try
@@ -50,7 +61,7 @@ namespace Savvyio
                     // by design; ignore potential orphaned dll references causing ReflectionTypeLoadException
                 }
             }
-            
+
             foreach (var contract in definedTypes.Where(type => type.IsInterface))
             {
                 var filtered = definedTypes.Where(type => SavvyioOptions.IsValid<T>(type, contract)).ToList();