+++ test/Savvyio.Core.Tests/Dispatchers/FireForgetDispatcherTest.cs
@@ -33,7 +33,7 @@ namespace Savvyio.Dispatchers
         public void Dispatch_ShouldDispatchCommandToDesignatedHandler()
         {
             var sc = new ServiceCollection()
-                .AddSavvyIO(o => o.AddAssemblyRangeToScan(typeof(FireForgetDispatcherTest).Assembly, typeof(ICommand).Assembly).EnableAutomaticDispatcherDiscovery().EnableAutomaticHandlerDiscovery())
+                .AddSavvyIO(o => o.AddAssemblyRangeToScan(typeof(FireForgetDispatcherTest).Assembly, typeof(ICommand).Assembly).EnableDispatcherDiscovery().EnableHandlerDiscovery())
                 .AddScoped<ITestStore<ICommand>, InMemoryTestStore<ICommand>>();
             
             var sp = sc.BuildServiceProvider();
@@ -62,7 +62,7 @@ namespace Savvyio.Dispatchers
         public async Task DispatchAsync_ShouldDispatchCommandToDesignatedHandlerAsync()
         {
             var sc = new ServiceCollection()
-                .AddSavvyIO(o => o.AddAssemblyRangeToScan(typeof(FireForgetDispatcherTest).Assembly, typeof(ICommand).Assembly).EnableAutomaticDispatcherDiscovery().EnableAutomaticHandlerDiscovery())
+                .AddSavvyIO(o => o.AddAssemblyRangeToScan(typeof(FireForgetDispatcherTest).Assembly, typeof(ICommand).Assembly).EnableDispatcherDiscovery().EnableHandlerDiscovery())
                 .AddScoped<ITestStore<ICommand>, InMemoryTestStore<ICommand>>();
             
             var sp = sc.BuildServiceProvider();