+++ test/Savvyio.Core.Tests/SavvyioOptionsExtensionsTest.cs
@@ -1,10 +1,12 @@
-﻿using Cuemon.Extensions.Xunit;
+﻿using System.Linq;
+using Cuemon.Extensions.Xunit;
 using Savvyio.Assets;
 using Savvyio.Assets.Domain;
 using Savvyio.Assets.Domain.Handlers;
 using Savvyio.Commands;
 using Savvyio.Domain;
 using Savvyio.EventDriven;
+using Savvyio.Extensions;
 using Savvyio.Queries;
 using Xunit;
 using Xunit.Abstractions;
@@ -45,6 +47,27 @@ namespace Savvyio
         }
 
         [Fact]
+        public void AddDispatchers_ShouldAutomaticallyAddDispatchersFromReferencedAssemblies()
+        {
+            var sut = new SavvyioOptions().UseAutomaticDispatcherDiscovery();
+
+            Assert.NotEmpty(sut.DispatcherServiceTypes);
+            Assert.NotEmpty(sut.DispatcherImplementationTypes);
+            Assert.Collection(sut.DispatcherServiceTypes.OrderBy(type => type.Name),
+                type => Assert.Equal(typeof(ICommandDispatcher), type),
+                type => Assert.Equal(typeof(IDomainEventDispatcher), type),
+                type => Assert.Equal(typeof(IIntegrationEventDispatcher), type),
+                type => Assert.Equal(typeof(IMediator), type),
+                type => Assert.Equal(typeof(IQueryDispatcher), type));
+            Assert.Collection(sut.DispatcherImplementationTypes.OrderBy(type => type.Name),
+                type => Assert.Equal(typeof(CommandDispatcher), type),
+                type => Assert.Equal(typeof(DomainEventDispatcher), type),
+                type => Assert.Equal(typeof(IntegrationEventDispatcher), type),
+                type => Assert.Equal(typeof(Mediator), type),
+                type => Assert.Equal(typeof(QueryDispatcher), type));
+        }
+
+        [Fact]
         public void AddHandlers_ShouldNotAddAnyHandlersFromCurrentDomain()
         {
             var sut = new SavvyioOptions().AddHandlers();
@@ -71,6 +94,29 @@ namespace Savvyio
                 type => Assert.Equal(typeof(PlatformProviderDomainEventHandler), type),
                 type => Assert.Equal(typeof(AccountEventHandler), type),
                 type => Assert.Equal(typeof(AccountQueryHandler), type));
+        }
+
+        [Fact]
+        public void AddHandlers_ShouldAddHandlersAutomaticallyFromReferencedAssemblies()
+        {
+            var sut = new SavvyioOptions().UseAutomaticHandlerDiscovery();
+
+            Assert.NotEmpty(sut.HandlerServiceTypes);
+            Assert.NotEmpty(sut.HandlerImplementationTypes);
+            Assert.Collection(sut.HandlerServiceTypes.OrderBy(type => type.Name),
+                type => Assert.Equal(typeof(ICommandHandler), type),
+                type => Assert.Equal(typeof(IDomainEventHandler), type),
+                type => Assert.Equal(typeof(IIntegrationEventHandler), type),
+                type => Assert.Equal(typeof(IQueryHandler), type));
+            Assert.Collection(sut.HandlerImplementationTypes.OrderBy(type => type.Name),
+                type => Assert.Equal(typeof(AccountCommandHandler), type),
+                type => Assert.Equal(typeof(AccountDomainEventHandler), type),
+                type => Assert.Equal(typeof(AccountEventHandler), type),
+                type => Assert.Equal(typeof(AccountHandlers), type),
+                type => Assert.Equal(typeof(AccountQueryHandler), type),
+                type => Assert.Equal(typeof(PlatformProviderDomainEventHandler), type),
+                type => Assert.Equal(typeof(PlatformProviderHandler), type)
+                );
         }
     }
 }